# Use WeakMap for caching