--- a/src/index.mjs
+++ b/src/index.mjs
@@ -11,14 +11,15 @@
  * limitations under the License.
  */
 
-const CACHE = {};
+const CACHE = new WeakMap();
 
 const stringify = JSON.stringify;
 
 export default function html(statics) {
-	let key = '.';
-	for (let i=0; i<statics.length; i++) key += statics[i].length + ',' + statics[i];
-	const tpl = CACHE[key] || (CACHE[key] = build(statics));
+	let tpl = CACHE.get(statics);
+	if (!tpl) {
+		CACHE.set(statics, tpl = build(statics));
+	}
 
 	// eslint-disable-next-line prefer-rest-params
 	return tpl(this, arguments);