@WebReflection 2018-12-23
Just 'cause you mentioned the conversation, all this has been tested for over a year through hyperHTML, and these are the things to consider:

  * IE 11 doesn't have template literals, but Babel 7 (and 6) preserve the standard behavior ("_slightly_" differently due [recent specs change](https://itnext.io/a-tiny-disastrous-ecmascript-change-fadc05c83e69))
  * IE < 11 doesn't have WeakMap (neither probably Map) but it has Babel freezing the template, meaning you need to monkey patch Babel upfront or common WeakMap polypill cannot work due frozen Array, something like this on top of the page would do: `this.WeakMap||(function(O,f){f=O.freeze||O;O.freeze=function(o){return'raw'in o?o:f(o)}}(Object));`
  * with that, you can bring WeakMaps down to any IE and legacy browser using [@ungap/weakmap](https://github.com/ungap/weakmap)
  * **TypeScript is broken**, and so is Firefox < 55, so you may want, in case you are not transpiling via Babel 6/7, fix at runtime the template literal uniqueness. I use [@ungap/template-literal](https://github.com/ungap/template-literal) for that, which is battle tested

If you ditch the WeakMap module, 'cause nowadays pretty much everyone targets IE 11 as baseline, where WeakMap is already available, you most likely need to address the uniqueness of the template literal issue due bad transpilers such TypeScript (2 versions with two different bugs 'cause they can't get specs right, maybe latter works).

If you think this is worth the performance improvement over standard browsers/transpiled code, you also have the module to include 😉 

If you think this project should never ship in production without pre-processing, then maybe you might also ignore the WeakMap/Template improvement, and focus on avoiding CSP issues due code evaluation.

I hope I've properly explained template literals gotchas, since older projects have already dealt with all of these to grant universal compatibility through any client/server engine.

Regards.

P.S. if size is that important, maybe this is a quick win too
```js
function html(statics) {
	return '.' + statics.map(c => [c.length, c]).join('');
}
```