@developit 2018-12-23
Could the produced function be cached as a property set on the statics array?