@WebReflection 2018-12-23
Nope, 'cause template literals are frozen arrays by specs