@jviide 2018-12-24
Thanks for the good info @WebReflection. It certainly sounds like experience earned through hard work :) Your thoughts, @developit - is WeakMap caching worth the trouble?

The CSP point you raised is a very valid one. PR #53 is an attempt to address that.